requirements.txt
defusedxml==0.5.0
Django==1.11
django-tastypie==0.14.1
lxml==4.2.1
pkg-resources==0.0.0
python-dateutil==2.7.3
python-mimeparse==1.6.0
pytz==2018.4
six==1.11.0
